// src/http-busca-search/models/modelsFilteredCategoryResult.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Mecanismo de Busca API
 * API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini
 * OpenAPI spec version: 1.0
 */
import type { ModelsServiceDocument } from './modelsServiceDocument'

export interface ModelsFilteredCategoryResult {
  name?: string
  page?: number
  per_page?: number
  services?: ModelsServiceDocument[]
  total_services?: number
}