// src/http-busca-search/models/modelsSubcategoryResponse.ts
import type { ModelsSubcategory } from './modelsSubcategory'
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Mecanismo de Busca API
 * API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini
 * OpenAPI spec version: 1.0
 */
import type { ModelsSubcategoryResponseMetadata } from './modelsSubcategoryResponseMetadata'

export interface ModelsSubcategoryResponse {
  category?: string
  metadata?: ModelsSubcategoryResponseMetadata
  subcategories?: ModelsSubcategory[]
  total_subcategories?: number
}