// src/http-busca-search/models/modelsSubcategoryServicesResponse.ts
import type { ModelsServiceDocument } from './modelsServiceDocument'
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Mecanismo de Busca API
 * API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini
 * OpenAPI spec version: 1.0
 */
import type { ModelsSubcategoryServicesResponseMetadata } from './modelsSubcategoryServicesResponseMetadata'

export interface ModelsSubcategoryServicesResponse {
  metadata?: ModelsSubcategoryServicesResponseMetadata
  page?: number
  per_page?: number
  services?: ModelsServiceDocument[]
  subcategory?: string
  total_services?: number
}