// src/http-pref-rio-carta-servicos/models/serviceDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API Carta de Serviços - Pref.Rio
 * API pública para consumo da Carta de Serviços Municipais do Rio de Janeiro, integrada com Salesforce.
 * OpenAPI spec version: 1.0.1
 */
import type { Button } from './button'
import type { Channel } from './channel'
import type { Legislation } from './legislation'
import type { ServiceDetailArticleStatus } from './serviceDetailArticleStatus'
import type { ServiceDetailFlags } from './serviceDetailFlags'
import type { ServiceDetailHowToRequest } from './serviceDetailHowToRequest'
import type { ServiceDetailInfo } from './serviceDetailInfo'
import type { ServiceDetailServicePointsItem } from './serviceDetailServicePointsItem'

export interface ServiceDetail {
  slug?: string
  name?: string
  serviceCatalogId?: string
  articleId?: string
  articleVersion?: number
  articleStatus?: ServiceDetailArticleStatus
  createdDate?: string
  lastModifiedDate?: string
  lastPublishedDate?: string
  slugHistory?: string[]
  themeSlug?: string
  themeName?: string
  subthemeSlug?: string
  subthemeName?: string
  responsibleOrgUnit?: string
  responsibleOrgUnitShortName?: string
  flags?: ServiceDetailFlags
  info?: ServiceDetailInfo
  howToRequest?: ServiceDetailHowToRequest
  channels?: Channel[]
  buttons?: Button[]
  legislation?: Legislation[]
  servicePoints?: ServiceDetailServicePointsItem[]
}