// src/http-pref-rio-carta-servicos/models/themesResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API Carta de Serviços - Pref.Rio
 * API pública para consumo da Carta de Serviços Municipais do Rio de Janeiro, integrada com Salesforce.
 * OpenAPI spec version: 1.0.1
 */
import type { PaginationMeta } from './paginationMeta'
import type { Theme } from './theme'

export interface ThemesResponse {
  data?: Theme[]
  meta?: PaginationMeta
}